ai-content-quiz-engine-main/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-dotenv
pymupdf
google-generativeai